Create ArrayLeaders.java                   Difficulty : Easy
link : https://www.geeksforgeeks.org/problems/leaders-in-an-array-1587115620/1

Description : You are given an array arr of positive integers. Your task is to find all the leaders in the array. An element is considered a leader if it is greater than or equal to all elements to its right. The rightmost element is always a leader.


Constraints:
1 <= arr.size() <= 106
0 <= arr[i] <= 106

Company Tags : 
Payu,  Adobe