Create EquilibriumPoint.java        Difficulty : Easy
link : https://www.geeksforgeeks.org/problems/equilibrium-point-1587115620/1

description : Given an array of integers arr[], the task is to find the first equilibrium point in the array.

The equilibrium point in an array is an index (0-based indexing) such that the sum of all elements before that index is the same as the sum of elements after it. Return -1 if no such point exists. 

Company Tags : 
Amazon Adobe Google Facebook Microsoft Uber

Topic Tags
prefix-sum, Arrays, Data Structures Algorithms