Create MeanAndMedian.java                    Difficulty : Easy
Given an array a[ ] of size N. The task is to find the median and mean of the array elements. Mean is average of the numbers and median is the element which is smaller than half of the elements and greater than remaining half.  If there are odd elements, the median is simply the middle element in the sorted array. If there are even elements, then the median is floor of average of two middle numbers in the sorted array. If mean is floating point number, then we need to print floor of it.

Note: To find the median, you might need to sort the array. Since sorting is covered in later tracks, we have already provided the sort function to you in the code.

Your Task:
You don't need to read input or print anything.. You just need to complete the following two function:

mean(): It takes the array and its size N as parameters and returns the mean as an integer.
median(): It takes the array and its size N as parameters and returns the median as an integer.

Expected Time Complexity: O(N log(N)).
Expected Auxiliary Space: O(1).

Constraints:
1 <= N <= 10^6
1 <= a[i] <= 10^6

Topic Tags
Mathematical, Algorithms