Create LargestFloor.java                             Difficulty : Easy
link : https://www.geeksforgeeks.org/problems/floor-in-a-sorted-array-1587115620/1

description : 
Given a sorted array arr[] (with unique elements) and an integer k, find the index (0-based) of the largest element in arr[] that is less than or equal to k. This element is called the "floor" of k. If such an element does not exist, return -1.

Constraints:
1 ≤ arr.size() ≤ 10^6
1 ≤ arr[i] ≤ 10^6
0 ≤ k ≤ arr[n-1]

Company Tags
Amazon